Move LightCurve class
