+++ gammapy/catalog/fermi.py
@@ -15,8 +15,7 @@ from gammapy.modeling.models import (
     SkyModel,
     TemplateSpatialModel,
 )
-from gammapy.estimators import FluxPoints
-from gammapy.time import LightCurve
+from gammapy.estimators import FluxPoints, LightCurve
 from gammapy.utils.gauss import Gauss2DPDF
 from gammapy.utils.scripts import make_path
 from gammapy.utils.table import table_standardise_units_inplace
